+++ src/Select.jsx
@@ -1,10 +1,12 @@
 import React, { Component, PropTypes, Children, cloneElement } from 'react'
 import ReactDOM, { findDOMNode } from 'react-dom'
+import { Select as ReactARIASelect } from 'react-aria'
 import TetherComponent from 'react-tether'
-import EventsHandler from './Events-Handler'
+import Measure from 'react-measure'
+import tabbable from 'tabbable'
 import childrenPropType from './children-prop-type'
 
-const eventsHandler = new EventsHandler()
+const { Manager } = ReactARIASelect
 
 class Select extends Component {
   static propTypes = {
@@ -14,7 +16,6 @@ class Select extends Component {
     offset: PropTypes.string,
     classPrefix: PropTypes.string,
     autoWidth: PropTypes.bool,
-    renderOptions: PropTypes.func,
     onChange: PropTypes.func,
     children: childrenPropType
   }
@@ -26,13 +27,7 @@ class Select extends Component {
     offset: '0px 0px',
     classPrefix: 'selectly',
     autoWidth: true,
-    renderOptions: options => options,
     onChange: () => null
-  }
-
-  static childContextTypes = {
-    onOptionSelect: PropTypes.func,
-    isOpen: PropTypes.bool
   }
 
   state = {
@@ -40,36 +35,11 @@ class Select extends Component {
     width: null
   }
 
-  getChildContext() {
-    return {
-      onOptionSelect: option => {
-        if (!this.props.multiple) {
-          this.setOpen(false)
-        }
-        this.props.onChange(option)
-      },
-      isOpen: this.state.isOpen
-    }
-  }
-
-  componentDidMount() {
-    // set the tethered content width
-    this._setWidth()
-
-    // add component to events handler so we delegate everything to one handler
-    // rather than every component instance
-    eventsHandler.add(this)
-  }
-
   componentWillReceiveProps(nextProps) {
     if (this.props.disabled !== nextProps.disabled &&
         nextProps.disabled === true) {
       this.setOpen(false)
     }
-  }
-
-  componentWillUnmount() {
-    eventsHandler.remove(this)
   }
 
   setOpen(isOpen, cb = () => null) {
@@ -80,78 +50,92 @@ class Select extends Component {
     return this.state.isOpen
   }
 
-  _setWidth() {
+  _handleMeasure = ({ width }) => {
     if (this.props.autoWidth) {
-      this.setState({width: this._trigger.offsetWidth})
+      this.setState({ width })
     }
   }
 
-  _toggleOpen({ target }) {
+  _handleToggle(firstChild, e) {
     if (this.props.disabled) return
 
-    if (this._trigger.contains(target)) {
-      this.setOpen(!this.state.isOpen)
-    } else if (this._options && !this._options.contains(target)) {
+    this.setOpen(!this.state.isOpen)
+
+    if (typeof firstChild.props.onTrigger === 'function') {
+      firstChild.props.onTrigger(e)
+    }
+  }
+
+  _handleTriggerKeyDown = ({ key }) => {
+    // determine if we need to move focus to the options menu when pressing tab
+    // while the menu is open
+    if (key === 'Tab' && this.state.isOpen) {
+      // without setTimeout it will focus the second tabbable item, need to figure
+      // out why this is happening
+      setTimeout(() => {
+        const tabbableChildren = tabbable(this._options)
+        if (tabbableChildren.length) {
+          tabbableChildren[0].focus()
+        }
+      })
+    }
+  }
+
+  _handleOptionSelection(secondChild, option, event) {
+    if (!this.props.multiple) {
       this.setOpen(false)
     }
-  }
 
-  _handleFocus = ({ key }) => {
-  }
-
-  _handleBlur = () => {
-    this.setOpen(false)
-  }
-
-  _handleKeyDown = ({ key }) => {
-    if (key === 'ArrowDown') {
-      this.setOpen(true)
-      this._options.focus()
+    if (typeof secondChild.props.onOptionSelection === 'function') {
+      secondChild.props.onOptionSelection(option, event)
     }
   }
 
-  _getKeyboardControls() {
-    return {
-      onFocus: this._handleFocus,
-      onBlur: this._handleBlur,
-      onKeyDown: this._handleKeyDown
+  _handleRequestClose(secondChild) {
+    this.setOpen(false)
+
+    if (typeof secondChild.props.onRequestClose === 'function') {
+      secondChild.props.onRequestClose()
     }
   }
 
   render() {
-    const { offset, classPrefix, autoWidth, children, renderOptions } = this.props
+    const { offset, classPrefix, autoWidth, children } = this.props
     const { isOpen, width } = this.state
-    const childrenArray = Children.toArray(children)
-    const firstChild = childrenArray[0]
-    const secondChild = childrenArray[1]
-
+    const [firstChild, secondChild] = Children.toArray(children)
     return (
-      <TetherComponent
-        attachment="top left"
-        targetAttachment="bottom left"
-        offset={offset}
-        classPrefix={classPrefix}
-        constraints={[{
-          to: 'window',
-          attachment: 'together'
-        }]}
-        style={{
-          width: width ? width + 'px' : ''
-        }}
-      >
-        { cloneElement(firstChild, {
-            ref: c => { this._trigger = findDOMNode(c) },
-            keyboardControls: this._getKeyboardControls()
-          })
-        }
-        { renderOptions(
-            isOpen &&
+      <Manager>
+        <TetherComponent
+          attachment="top left"
+          targetAttachment="bottom left"
+          offset={offset}
+          classPrefix={classPrefix}
+          constraints={[{
+            to: 'window',
+            attachment: 'together'
+          }]}
+          style={{ width: width ? `${width}px` : '' }}
+        >
+          <Measure onMeasure={this._handleMeasure}>
+            { cloneElement(firstChild, {
+                ref: c => this._trigger = findDOMNode(c),
+                isOpen,
+                keybindings: isOpen ? [] : [' '],
+                onTrigger: this._handleToggle.bind(this, firstChild),
+                onKeyDown: this._handleTriggerKeyDown
+              })
+            }
+          </Measure>
+          { isOpen &&
             cloneElement(secondChild, {
-              ref: c => { this._options = findDOMNode(c) }
+              ref: c => this._options = findDOMNode(c),
+              closeOnOutsideClick: true,
+              onOptionSelection: this._handleOptionSelection.bind(this, secondChild),
+              onRequestClose: this._handleRequestClose.bind(this, secondChild)
             })
-          )
-        }
-      </TetherComponent>
+          }
+        </TetherComponent>
+      </Manager>
     )
   }
 }