+++ src/selectly.js
@@ -1,11 +1,12 @@
-export { default as Select } from './Select'
-export { default as Option } from './Option'
-
+import { Trigger, Select as ReactARIASelect } from 'react-aria'
+import Select from './Select'
 import buildOptionsLookup from './utils/build-options-lookup'
 import getCurrentOptions from './utils/get-current-options'
 import getToggledOptions from './utils/get-toggled-options'
 import getAllValues from './utils/get-all-values'
 import isOptionSelected from './utils/is-option-selected'
+
+const { OptionList, Option } = ReactARIASelect
 
 export const utils = {
   buildOptionsLookup,
@@ -13,4 +14,11 @@ export const utils = {
   getToggledOptions,
   getAllValues,
   isOptionSelected
+}
+
+export {
+  Select,
+  Trigger,
+  OptionList,
+  Option
 }